% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{births_wpp_2019}
\alias{births_wpp_2019}
\title{The Philippines births by age group of mother for 2020 from the World
Population Prospects 2019}
\format{
A tibble with 4 columns and 21 rows:
\describe{
  \item{\code{area}}{Area i.e., country}
  \item{\code{year}}{Year i.e., 2015-2020}
  \item{\code{age_category}}{Five-year age groups}
  \item{\code{birth}}{Number of births}
}
}
\source{
World Population Prospects 2019
}
\usage{
births_wpp_2019
}
\description{
The Philippines births by age group of mother for 2020 from the World
Population Prospects 2019
}
\examples{
births_wpp_2019


}
\keyword{datasets}
