% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comoparams.R
\docType{package}
\name{comoparams}
\alias{comoparams}
\alias{comoparams-package}
\title{R Utility Tool to Extract, Process and Structure Philippines-specific Datasets
for Use in CoMo Consortium Model}
\description{
The Oxford Modelling Group for Global Health (OMGH) is developing model
structures to estimate the impact of potential mitigation strategies. By
changing the input data and parameter values, the models can be adjusted to
represent a national or subnational setting. In addition, a user friendly
application and interface is being developed to enable widespread utility.
This package supports in the processing of respective national and/or
subnational input data and parameter values into data structures appropriate
for use with the modelling application and/or other packages developed for
this purpose.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/como-ph/comoparams}
  \item Report bugs at \url{https://github.com/como-ph/comoparams/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

Authors:
\itemize{
  \item Chris Mercado
  \item Robert Medina
}

}
\keyword{internal}
