% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataSources}
\alias{dataSources}
\title{Table of data sources used for parameters}
\format{
A tibble with 5 columns and 8 rows:
\describe{
  \item{\code{description}}{Description of data source}
  \item{\code{parameter}}{Model parameter data source addresses}
  \item{\code{source}}{Issuing authority or body of the data source}
  \item{\code{scode}}{Source code}
  \item{\code{format}}{Format of the data available from the data source}
  \item{\code{link}}{URL for data source}
}
}
\usage{
dataSources
}
\description{
Table of data sources used for parameters
}
\examples{
dataSources


}
\keyword{datasets}
