% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-get_como_template.R
\name{get_como_template}
\alias{get_como_template}
\title{Retrieve CoMo parameters template}
\usage{
get_como_template()
}
\value{
A list of 9 elements for each of the worksheets in the CoMo
  parameters template
}
\description{
Retrieve CoMo parameters template
}
\examples{
get_como_template()

}
