% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ph_calculate_rates.R
\name{ph_calculate_nus}
\alias{ph_calculate_nus}
\title{Duration of hospitalised infection}
\usage{
ph_calculate_nus(df)
}
\arguments{
\item{df}{A data.frame of cases data pulled from the Philippines COVID-19
Data Drop.}
}
\value{
Value for duration of hospitalised infection.
}
\description{
Duration of hospitalised infection
}
\examples{
df <- ph_get_cases()
ph_calculate_nus(df = df)

}
