% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ph_calculate_rates.R
\name{ph_calculate_pdeath}
\alias{ph_calculate_pdeath}
\title{Calculate probability of dying when hospitalised}
\usage{
ph_calculate_pdeath(df)
}
\arguments{
\item{df}{A data.frame of cases data pulled from the Philippines COVID-19
Data Drop.}
}
\value{
Value for probability of dying when hospitalised.
}
\description{
Calculate probability of dying when hospitalised
}
\examples{
df <- ph_get_cases()
ph_calculate_pdeath(df = df)

}
