% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ph_calculate_rates.R
\name{ph_calculate_rates}
\alias{ph_calculate_rates}
\title{Calculate infection fatality rate (IFR), infection hospitalisation rate, and
hospitalisation fatality rate (HFR) from Philippines data}
\usage{
ph_calculate_rates(df)
}
\arguments{
\item{df}{A data.frame of cases data pulled from the Philippines COVID-19
Data Drop.}
}
\value{
A data.frame with calculated IHR, IFR, and HFR by age group structure
  consistent with CoMo model data structure requirements.
}
\description{
Calculate infection fatality rate (IFR), infection hospitalisation rate, and
hospitalisation fatality rate (HFR) from Philippines data
}
\examples{
df <- ph_get_cases()
ph_calculate_rates(df = df)

}
