% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/23-ph_create_params.R
\name{ph_create_params}
\alias{ph_create_params}
\title{Output cases and severity-mortality input tables compatible with CoMo
template}
\usage{
ph_create_params(params, path = ".")
}
\arguments{
\item{params}{A list of parameters produced using the \code{ph_set_params()}
function}

\item{path}{Directory to save output to. Defaults to current working
directory}
}
\value{
An XLSX workbook with 10 worksheets for various sets of parameters
  formatted according to the CoMo template
}
\description{
Output cases and severity-mortality input tables compatible with CoMo
template
}
\examples{
\dontrun{
  ph_create_params()
}

}
