% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ph_datadrop.R
\name{ph_gdrive_files}
\alias{ph_gdrive_files}
\title{Get Department of Health COVID-19 Data Drop Google Drive specific directory
information}
\usage{
ph_gdrive_files(version = "current", date = NULL)
}
\arguments{
\item{version}{A character value specifying whether to get the most
currently available dataset (\code{"current"}) or to get archive data
(\code{"archive"}). Default to \code{"current"}}

\item{date}{Date in <YYYY-MM-DD> format; This is the date up to which
extracted data reports to. Only used when \code{version} is set to
\code{"archive"} otherwise ignored.}
}
\value{
A tibble containing information on the various files in a specified
  COVID-19 Data Drop Google Drive directory
}
\description{
A wrapper to googledrive functions to extract information on contents of
a specific COVID-19 Data Drop Google Drive directory
}
\examples{
ph_gdrive_files()

}
