% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_datadrop.R
\name{ph_get_fields}
\alias{ph_get_fields}
\title{Pull Philippines data fields information from the publicly available
Department of Health COVID-19 Data Drop}
\usage{
ph_get_fields()
}
\value{
A tibble
}
\description{
A wrapper to the googlesheets4 read_sheet() function to pull data from the
Philppines' COVID-19 Data Drop resource that is publicly distributed via
Google Sheets
}
\examples{
ph_get_fields()

}
