% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ph_get_population.R
\name{ph_get_popcen2015}
\alias{ph_get_popcen2015}
\title{Get Philippines population estimates for 2015 and projections for 2016-2025
based on the results of the Philippines 2015 POPCEN}
\usage{
ph_get_popcen2015(file)
}
\arguments{
\item{file}{Either a path or a URL to the 2015 POPCEN XLSX file containing
population projections}
}
\value{
A tibble in tidy format containing population data by 5-year age
  groups for the entire Philippines and by region from 2015 (estimates) to
  2025 (projections)
}
\description{
Get Philippines population estimates for 2015 and projections for 2016-2025
based on the results of the Philippines 2015 POPCEN
}
\examples{
link <- "https://psa.gov.ph/sites/default/files/attachments/hsd/pressrelease/"
fname <- "Updated\%20Population\%20Projections\%20based\%20on\%202015\%20POPCEN_0.xlsx"
ph_get_popcen2015(file = paste(link, fname, sep = ""))

}
