% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ph_get_population.R
\name{ph_get_wpp2019}
\alias{ph_get_wpp2019}
\title{Get Philippines population estimates or projections by 5-year age groups
given a year or range of years from the World Population Prospects 2019}
\usage{
ph_get_wpp2019(
  file,
  location = "Philippines",
  year = lubridate::year(Sys.Date())
)
}
\arguments{
\item{file}{Either a path or a URL to the World Population Prospects 2019
file containing population estimates and projections in 5-year age groups}

\item{location}{Location to get population estimates/projections for;
Default to the "Philippines"}

\item{year}{A year (numeric) or range of years (YYYY:YYYY) to get population
estimates/projections for; Default to current year.}
}
\value{
A tibble in tidy format containing population data by 5-year age
  groups for the entire Philippines and by region from 2015 (estimates) to
  2025 (projections)
}
\description{
Get Philippines population estimates or projections by 5-year age groups
given a year or range of years from the World Population Prospects 2019
}
\examples{
link <- "https://population.un.org/wpp/Download/Files/1_Indicators\%20(Standard)/CSV_FILES/"
fname <- "WPP2019_PopulationByAgeSex_Medium.csv"
ph_get_wpp2019(file = paste(link, fname, sep = ""))

}
