% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ph_get_population.R
\name{ph_get_wpp2019_pop}
\alias{ph_get_wpp2019_pop}
\title{Get Philippines population estimates or projections by 5-year age groups
given a year or range of years from the World Population Prospects 2019}
\usage{
ph_get_wpp2019_pop(
  file,
  location = "Philippines",
  period = lubridate::year(Sys.Date())
)
}
\arguments{
\item{file}{Either a path or a URL to the World Population Prospects 2019
file containing population estimates and projections in 5-year age groups}

\item{location}{Location to get population estimates/projections for;
Default to the "Philippines"}

\item{period}{A year (numeric) or range of years (YYYY:YYYY) to get population
estimates/projections for; Default to current year.}
}
\value{
A tibble in tidy format containing population data by 5-year age
  groups for the entire Philippines and by region from 2015 (estimates) to
  2025 (projections)
}
\description{
Get Philippines population estimates or projections by 5-year age groups
given a year or range of years from the World Population Prospects 2019
}
\examples{
\dontrun{
link <- with(dataSources,
             link[source == "World Population Prospects 2019" &
             parameter == "population"])
ph_get_wpp2019_pop(file = link)
}

}
