% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-ph_set_cases.R
\name{ph_set_cases}
\alias{ph_set_cases}
\title{Set Philippines-specific CoMo model simulation parameters for COVID-19 cases}
\usage{
ph_set_cases()
}
\value{
A data.frame of Philippines-specific CoMo model parameters for
  COVID-19 cases
}
\description{
Set Philippines-specific CoMo model simulation parameters for COVID-19 cases
}
\examples{
ph_set_cases()

}
