% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17-ph_set_elderly.R
\name{ph_set_cocoon}
\alias{ph_set_cocoon}
\title{Set Philippines-specific CoMo model simulation for shielding the elderly
intervention parameters}
\usage{
ph_set_cocoon()
}
\value{
A list of Philippines-specific CoMo model shielding the elderly
intervention parameters
}
\description{
Set Philippines-specific CoMo model simulation for shielding the elderly
intervention parameters
}
\examples{
if(interactive()) ph_set_cocoon()

}
