% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13-ph_set_distancing.R
\name{ph_set_distancing}
\alias{ph_set_distancing}
\title{Set Philippines-specific CoMo model simulation social distancing intervention
parameters}
\usage{
ph_set_distancing()
}
\value{
A list of Philippines-specific CoMo model social distancing
intervention parameters
}
\description{
Set Philippines-specific CoMo model simulation social distancing intervention
parameters
}
\examples{
if(interactive()) ph_set_distancing()

}
