% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-ph_set_general.R
\name{ph_set_general}
\alias{ph_set_general}
\title{Set Philippines-specific CoMo model simulation parameters for general and
area factors}
\usage{
ph_set_general()
}
\value{
A list of Philippines-specific CoMo model parameters for general and
  area factors
}
\description{
Set Philippines-specific CoMo model simulation parameters for general and
area factors
}
\examples{
ph_set_general()

}
