% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14-ph_set_handwashing.R
\name{ph_set_handwashing}
\alias{ph_set_handwashing}
\title{Set Philippines-specific CoMo model simulation handwashing intervention
parameters}
\usage{
ph_set_handwashing()
}
\value{
A list of Philippines-specific CoMo model handwashing intervention
  parameters
}
\description{
Set Philippines-specific CoMo model simulation handwashing intervention
parameters
}
\examples{
if(interactive()) ph_set_handwashing()

}
