% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-ph_set_hospital.R
\name{ph_set_hospital}
\alias{ph_set_hospital}
\title{Set Philippines-specific CoMo model simulation parameters for hospitalisation
parameters}
\usage{
ph_set_hospital()
}
\value{
A list of Philippines-specific CoMo model parameters for
  hospitalisation parameters
}
\description{
Set Philippines-specific CoMo model simulation parameters for hospitalisation
parameters
}
\examples{
if(interactive()) ph_set_hospital()

}
