% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-ph_set_household.R
\name{ph_set_household}
\alias{ph_set_household}
\title{Set Philippines-specific CoMo model simulation self-isolation intervention
parameters}
\usage{
ph_set_household()
}
\value{
A list of Philippines-specific CoMo model self-isolation intervention
  parameters
}
\description{
Set Philippines-specific CoMo model simulation self-isolation intervention
parameters
}
\examples{
if(inreractive()) ph_set_household()

}
