% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21-ph_set_interventions.R
\name{ph_set_interventions}
\alias{ph_set_interventions}
\title{Set Philippines-specific CoMo model simulation intervention parameters}
\usage{
ph_set_interventions()
}
\value{
A list of Philippines-specific CoMo model intervention parameters
}
\description{
Set Philippines-specific CoMo model simulation intervention parameters
}
\examples{
if(interactive()) ph_set_interventions()

}
