% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-ph_set_lockdown.R
\name{ph_set_lockdown}
\alias{ph_set_lockdown}
\title{Set Philippines-specific CoMo model simulation lockdown intervention
parameters}
\usage{
ph_set_lockdown()
}
\value{
A list of Philippines-specific CoMo model lockdown intervention
parameters
}
\description{
Set Philippines-specific CoMo model simulation lockdown intervention
parameters
}
\examples{
if(interactive()) ph_set_lockdown()

}
