% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-ph_set_params.R
\name{ph_set_params}
\alias{ph_set_params}
\title{Interactive guidance for setting Philppines-specific CoMo model parameters}
\usage{
ph_set_params()
}
\value{
A list of Philippines-specific CoMo model parameters
}
\description{
Interactive guidance for setting Philppines-specific CoMo model parameters
}
\examples{
if(interactive()) ph_set_params()

}
