% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-ph_set_population.R
\name{ph_set_population}
\alias{ph_set_population}
\title{Set Philippines-specific CoMo model simulation parameters for population}
\usage{
ph_set_population()
}
\value{
A data.frame of Philippines-specific CoMo model parameters for
  population
}
\description{
Set Philippines-specific CoMo model simulation parameters for population
}
\examples{
ph_set_population()

}
