% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/19-ph_set_quarantine.R
\name{ph_set_quarantine}
\alias{ph_set_quarantine}
\title{Set Philippines-specific CoMo model simulation voluntary home quarantine
intervention parameters}
\usage{
ph_set_quarantine()
}
\value{
A list of Philippines-specific CoMo model voluntary home quarantine
intervention parameters
}
\description{
Set Philippines-specific CoMo model simulation voluntary home quarantine
intervention parameters
}
\examples{
if(interactive()) ph_set_quarantine()

}
