% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16-ph_set_school.R
\name{ph_set_school}
\alias{ph_set_school}
\title{Set Philippines-specific CoMo model simulation schools closure intervention
parameters}
\usage{
ph_set_school()
}
\value{
A list of Philippines-specific CoMo model school clostures
intervention parameters
}
\description{
Set Philippines-specific CoMo model simulation schools closure intervention
parameters
}
\examples{
if(interactive()) ph_set_school()

}
