% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-ph_set_screening.R
\name{ph_set_screening}
\alias{ph_set_screening}
\title{Set Philippines-specific CoMo model simulation screening intervention
parameters}
\usage{
ph_set_screening()
}
\value{
A list of Philippines-specific CoMo model screening intervention
  parameters
}
\description{
Set Philippines-specific CoMo model simulation screening intervention
parameters
}
\examples{
ph_set_screening()

}
