% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-ph_set_severe.R
\name{ph_set_severe}
\alias{ph_set_severe}
\title{Set Philippines-specific CoMo model simulation parameters for severity of
cases and mortality}
\usage{
ph_set_severe()
}
\value{
A data.frame of Philippines-specific CoMo model parameters for
  severity of cases and mortality
}
\description{
Set Philippines-specific CoMo model simulation parameters for severity of
cases and mortality
}
\examples{
if(interactive()) ph_set_severe()

}
