% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18-ph_set_travel.R
\name{ph_set_travel}
\alias{ph_set_travel}
\title{Set Philippines-specific CoMo model simulation travel ban intervention
parameters}
\usage{
ph_set_travel()
}
\value{
A list of Philippines-specific CoMo model travel ban intervention
parameters
}
\description{
Set Philippines-specific CoMo model simulation travel ban intervention
parameters
}
\examples{
if(interactive()) ph_set_travel()

}
