% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-ph_set_vaccination.R
\name{ph_set_vaccination}
\alias{ph_set_vaccination}
\title{Set Philippines-specific CoMo model simulation vaccination intervention
parameters}
\usage{
ph_set_vaccination()
}
\value{
A list of Philippines-specific CoMo model vaccination intervention
parameters
}
\description{
Set Philippines-specific CoMo model simulation vaccination intervention
parameters
}
\examples{
if(interactive()) ph_set_vaccination()

}
