% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-ph_set_virus.R
\name{ph_set_virus}
\alias{ph_set_virus}
\title{Set Philippines-specific CoMo model simulation parameters for virus
parameters}
\usage{
ph_set_virus()
}
\value{
A list of Philippines-specific CoMo model parameters for virus
  parameters
}
\description{
Set Philippines-specific CoMo model simulation parameters for virus
parameters
}
\examples{
if(interactive()) ph_set_virus()

}
