% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-ph_set_work.R
\name{ph_set_work}
\alias{ph_set_work}
\title{Set Philippines-specific CoMo model simulation work from home intervention
parameters}
\usage{
ph_set_work()
}
\value{
A list of Philippines-specific CoMo model work from home intervention
parameters
}
\description{
Set Philippines-specific CoMo model simulation work from home intervention
parameters
}
\examples{
if(interactive()) ph_set_work()

}
