% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_psa_2015}
\alias{population_psa_2015}
\title{The Philippines 2015 Census of Population (POPCEN 2015)}
\format{
A tibble with 6 columns and 87567 rows:
\describe{
  \item{\code{area}}{Area (i.e., country, region, municipality, city) in
    the Philippines}
  \item{\code{year}}{Year (2015 to 2024)}
  \item{\code{age_category}}{Five-year age groups}
  \item{\code{total}}{Total population}
  \item{\code{male}}{Male population}
  \item{\code{female}}{Female population}
}
}
\source{
Philippines Statistics Authority
}
\usage{
population_psa_2015
}
\description{
The Philippines 2015 Census of Population (POPCEN 2015)
}
\examples{
population_psa_2015


}
\keyword{datasets}
