% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_wpp_2019}
\alias{population_wpp_2019}
\title{The Philippines population for 2020 from the World Population Prospects 2019}
\format{
A tibble with 6 columns and 21 rows:
\describe{
  \item{\code{area}}{Area i.e., country}
  \item{\code{year}}{Year i.e., 2020}
  \item{\code{age_category}}{Five-year age groups}
  \item{\code{total}}{Total population}
  \item{\code{male}}{Male population}
  \item{\code{female}}{Female population}
}
}
\source{
World Population Prospects 2019
}
\usage{
population_wpp_2019
}
\description{
The Philippines population for 2020 from the World Population Prospects 2019
}
\examples{
population_wpp_2019


}
\keyword{datasets}
