% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-ph_calculate_cases.R
\name{ph_calculate_cases}
\alias{ph_calculate_cases}
\title{Calculate daily cases and daily deaths from Philippines COVID-19 Data Drop}
\usage{
ph_calculate_cases(date = Sys.Date())
}
\arguments{
\item{date}{Date in <YYYY-MM-DD> format; This is the date up to which
extracted data reports to. Default is current date (\code{Sys.Date()})}
}
\value{
A data.frame with calculated number of cases per and number of deaths
  per day with structure consistent with CoMo model data structure
  requirements.
}
\description{
Calculate daily cases and daily deaths from Philippines COVID-19 Data Drop
}
\examples{
ph_calculate_cases(date = "2020-04-17")

}
