% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-ph_calculate_cases.R
\name{ph_calculate_cases}
\alias{ph_calculate_cases}
\title{Calculate daily cases, daily deaths and daily recoveries from Philippines
COVID-19 Data Drop}
\usage{
ph_calculate_cases(df)
}
\arguments{
\item{df}{A data.frame of cases data pulled from the Philippines COVID-19
Data Drop.}
}
\value{
A tibble with calculated number of cases per day and number
  of deaths per day and number of recovered per day with structure consistent
  with CoMo model data structure requirements.
}
\description{
Calculate daily cases, daily deaths and daily recoveries from Philippines
COVID-19 Data Drop
}
\examples{
ph_calculate_cases(df = ph_get_cases())

}
