% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ph_calculate_rates.R
\name{ph_calculate_nus}
\alias{ph_calculate_nus}
\title{Duration of hospitalised infection}
\usage{
ph_calculate_nus(date = Sys.Date())
}
\arguments{
\item{date}{Date in <YYYY-MM-DD> format; This is the date up to which
extracted data reports to. Default is current date (\code{Sys.Date()})}
}
\value{
Value for duration of hospitalised infection.
}
\description{
Duration of hospitalised infection
}
\examples{
ph_calculate_nus(date = "2020-04-17")

}
