% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-ph_calculate_rates.R
\name{ph_calculate_pdeath}
\alias{ph_calculate_pdeath}
\title{Calculate probability of dying when hospitalised}
\usage{
ph_calculate_pdeath(date = Sys.Date())
}
\arguments{
\item{date}{Date in <YYYY-MM-DD> format; This is the date up to which
extracted data reports to. Default is current date (\code{Sys.Date()})}
}
\value{
Value for probability of dying when hospitalised.
}
\description{
Calculate probability of dying when hospitalised
}
\examples{
ph_calculate_pdeath(date = "2020-04-17")

}
