% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_calculate_ifr.R
\name{ph_calculate_rates}
\alias{ph_calculate_rates}
\title{Calculate infection fatality rate (IFR) and infection hospitalisation rate
from Philippines data}
\usage{
ph_calculate_rates(
  date = stringr::str_remove_all(string = Sys.Date(), pattern = "-")
)
}
\arguments{
\item{date}{Date in format used by COVID-19 Data Drop system <YYYYMMDD>}
}
\value{
A data.frame with calculated IHR and IFR by age group
}
\description{
Calculate infection fatality rate (IFR) and infection hospitalisation rate
from Philippines data
}
