% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_calculate_rates.R
\name{ph_calculate_rates}
\alias{ph_calculate_rates}
\title{Calculate infection fatality rate (IFR), infection hospitalisation rate, and
hospitalisation fatality rate (HFR) from Philippines data}
\usage{
ph_calculate_rates(date = Sys.Date())
}
\arguments{
\item{date}{Date in <YYYY-MM-DD> format; This is the date up to which
extracted data reports to. Default is current date (\code{Sys.Date()})}
}
\value{
A data.frame with calculated IHR, IFR, and HFR by age group structure
  consistent with CoMo model data structure requirements.
}
\description{
Calculate infection fatality rate (IFR), infection hospitalisation rate, and
hospitalisation fatality rate (HFR) from Philippines data
}
\examples{
ph_calculate_rates(date = "2020-04-17")

}
