% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_calculate_ifr.R
\name{ph_calculate_rates}
\alias{ph_calculate_rates}
\title{Calculate infection fatality rate (IFR) and infection hospitalisation rate
from Philippines data}
\usage{
ph_calculate_rates(df = ph_get_cases())
}
\arguments{
\item{df}{Data.frame containing data on cases and deaths by age group.
Default is data.frame pulled from DoH DataDrop pulled using
\code{ph_get_cases()}}
}
\value{
A data.frame with calculated IHR and IFR by age group
}
\description{
Calculate infection fatality rate (IFR) and infection hospitalisation rate
from Philippines data
}
\examples{
ph_calculate_rates()

}
