% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-ph_create_params.R
\name{ph_create_params}
\alias{ph_create_params}
\title{Output cases and severity-mortality input tables compatible with CoMo
template}
\usage{
ph_create_params(date = Sys.Date(), path = "")
}
\arguments{
\item{date}{Date in <YYYY-MM-DD> format; This is the date up to which
extracted data reports to. Default is current date (\code{Sys.Date()})}

\item{path}{Directory to save output to}
}
\value{
An XLSX workbook with 2 worksheets for cases and severity-mortality
  formatted according to the CoMo template
}
\description{
Output cases and severity-mortality input tables compatible with CoMo
template
}
\examples{
\dontrun{
  ph_create_params()
}

}
