% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ph_datadrop.R
\name{ph_gdrive_files}
\alias{ph_gdrive_files}
\title{Get Department of Health COVID-19 Data Drop Google Drive specific directory
information}
\usage{
ph_gdrive_files(date = Sys.Date())
}
\arguments{
\item{date}{Date in <YYYY-MM-DD> format; This is the date up to which
extracted data reports to. Default is current date (\code{Sys.Date()}).
Data Drop started on 14 April 2020 so earliest date value that can be used
is 2020-04-14.}
}
\value{
A tibble containing information on the various files in a specified
  COVID-19 Data Drop Google Drive directory
}
\description{
A wrapper to googledrive functions to extract information on contents of
a specific COVID-19 Data Drop Google Drive directory
}
\examples{
ph_gdrive_files(date = "2020-06-01")

}
