% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_datadrop.R
\name{ph_get_cases}
\alias{ph_get_cases}
\title{Pull Philippines data on cases information from the publicly available
Department of Health COVID-19 Data Drop}
\usage{
ph_get_cases(
  date = stringr::str_remove_all(string = Sys.Date(), pattern = "-")
)
}
\arguments{
\item{date}{Date in format used by COVID-19 Data Drop system <YYYYMMDD>}
}
\value{
A tibble
}
\description{
A wrapper to googledrive and googlesheets4 functions to pull data from the
Philppines' COVID-19 Data Drop resource that is publicly distributed via
Google Drive
}
\examples{
ph_get_cases(date = "20200416")

}
