% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ph_datadrop.R
\name{ph_get_cases}
\alias{ph_get_cases}
\title{Pull Philippines data on cases information from the publicly available
Department of Health COVID-19 Data Drop}
\usage{
ph_get_cases(version = "current", date = NULL)
}
\arguments{
\item{version}{A character value specifying whether to get the most
currently available dataset (\code{"current"}) or to get archive data
(\code{"archive"}). Default to \code{"current"}}

\item{date}{Date in <YYYY-MM-DD> format; This is the date up to which
extracted data reports to. Only used when \code{version} is set to
\code{"archive"} otherwise ignored.}
}
\value{
A tibble of case information on confirmed COVID-19 cases distributed
  via the COVID-19 Data Drop
}
\description{
A wrapper to googledrive and googlesheets4 functions to pull data from the
Philppines' COVID-19 Data Drop resource that is publicly distributed via
Google Drive
}
\examples{
ph_get_cases()

}
