% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_datadrop.R
\name{ph_get_fields}
\alias{ph_get_fields}
\title{Pull Philippines data fields information from the publicly available
Department of Health COVID-19 Data Drop}
\usage{
ph_get_fields(date = Sys.Date())
}
\arguments{
\item{date}{Date in <YYYY-MM-DD> format; This is the date up to which
extracted data reports to. Default is current date (\code{Sys.Date()})}
}
\value{
A tibble
}
\description{
A wrapper to googledrive and googlesheets4 functions to pull data from the
Philppines' COVID-19 Data Drop resource that is publicly distributed via
Google Drive
}
\examples{
ph_get_fields(date = "2020-04-17")

}
