% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-ph_get_population.R
\name{ph_get_wpp2019_births}
\alias{ph_get_wpp2019_births}
\title{Get Philippines age-specific number of births estimates or projections
by 5-year periods given a year or range of years from the World Population
Prospects 2019}
\usage{
ph_get_wpp2019_births(
  file,
  location = "Philippines",
  period = lubridate::year(Sys.Date())
)
}
\arguments{
\item{file}{Either a path or a URL to the World Population Prospects 2019
file containing population estimates and projections in 5-year age groups}

\item{location}{Location to get population estimates/projections for;
Default to the "Philippines"}

\item{period}{A year (numeric) or range of years (YYYY:YYYY) in 5 year
intervals starting from 1950 to get population estimates/projections for;
Default is current year corresponding to the five year period that
contains the current year}
}
\value{
A tibble in tidy format containing age-specific number of births
  for the entire Philippines from 1950 to 2100
}
\description{
Get Philippines age-specific number of births estimates or projections
by 5-year periods given a year or range of years from the World Population
Prospects 2019
}
\examples{
link <- "https://population.un.org/wpp/Download/Files/1_Indicators\%20(Standard)/"
fname <- "EXCEL_FILES/2_Fertility/WPP2019_FERT_F06_BIRTHS_BY_AGE_OF_MOTHER.xlsx"
ph_get_wpp2019_births(file = paste(link, fname, sep = ""))

}
