% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-ph_set_interventions.R,
%   R/20-ph_set_vaccination.R
\name{ph_set_intervention}
\alias{ph_set_intervention}
\title{Set Philippines-specific CoMo model simulation intervention parameters}
\usage{
ph_set_intervention()

ph_set_intervention()
}
\value{
A list of Philippines-specific CoMo model intervention parameters

A list of Philippines-specific CoMo model intervention parameters
}
\description{
Set Philippines-specific CoMo model simulation intervention parameters

Set Philippines-specific CoMo model simulation intervention parameters
}
\examples{
if(interactive()) ph_set_interevention()

if(interactive()) ph_set_interevention()

}
